requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.24.3
Werkzeug==2.3.7 